+++ styles.css
@@ -177,7 +177,6 @@ a {
 
   #calendar-container div {
     padding: 10px 0;
-
   }
 
   #calendar-container {
